using updated order info ref, excel exporting properly with named ranges

+++ SaturnCalculator.entity/Models/CalculatedLineItems.cs
@@ -6,5 +6,5 @@ public class CalculatedLineItems
     public string VendorItemNumber { get; set; }
     public string ShippedQuantity { get; set; }
     public string ItemCost { get; set; }
-    public float TotalCost { get; set; }
+    public double TotalCost { get; set; }
 }
