+++ SaturnCalculator.entity/Models/CalculatedVendorItem.cs
@@ -5,5 +5,5 @@ public class CalculatedVendorItem
     public string VendorItemNumber { get; set; }
     public int TotalQuantity { get; set; }
     public int TotalWeight { get; set; }
-    public float TotalCost { get; set; }
+    public double TotalCost { get; set; }
 }
